requirements.txt
asgiref==3.5.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
flexicon==0.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
phonenumberslite==8.13.0
Pillow==9.3.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.2
simplejson==3.19.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.2.0
